requirements.txt
typer==0.12.3
fastapi==0.111.0
uvicorn[standard]==0.30.1
jinja2==3.1.4
pydantic==2.7.4
pyyaml==6.0.1
pytest==8.2.0
httpx==0.27.0
openai==0.28.0
jsonschema==4.22.0
sqlalchemy==2.0.31
black==24.4.2
ruff==0.5.5
mypy==1.10.0
